requirements.txt
streamlit
pytesseract
pillow
fpdf2